split NativeElement into Native and NativeView
