+++ src/dom/index.ts
@@ -1,16 +1,18 @@
 import { registerSvelteElements } from './svelte-elements'
 import { registerNativeElements } from './nativescript-elements'
 import SvelteNativeDocument from './svelte/SvelteNativeDocument'
-import NativeElementNode from './native/NativeElementNode'
+import NativeViewElementNode from './native/NativeViewElementNode'
 import { write, messageType } from 'tns-core-modules/trace'
 import { logger, LogLevel } from './basicdom'
+import { View } from 'tns-core-modules/ui/core/view'
 
 export { default as FrameElement } from "./native/FrameElement"
 export { default as SvelteNativeDocument } from './svelte/SvelteNativeDocument'
 export { default as NativeElementNode } from './native/NativeElementNode'
+export { default as NativeViewElementNode } from './native/NativeViewElementNode'
 export { default as TabsElement } from './native/TabsElement'
 export { default as TabStripElement } from './native/TabStripElement'
-export { registerElement, createElement, ViewNode } from './basicdom'
+export { registerElement, createElement, ViewNode, ElementNode, logger, LogLevel } from './basicdom'
 export { navigate, goBack, showModal, closeModal, ShowModalOptions, NavigationOptions, BackNavigationOptions } from './navigation'
 
 
@@ -26,7 +28,7 @@ function installGlobalShims(): SvelteNativeDocument {
         setTimeout(action, 33); //about 30 fps
     }
 
-    window.getComputedStyle = (node: NativeElementNode) => {
+    window.getComputedStyle = (node: NativeViewElementNode<View>) => {
         return node.nativeView.style;
     }
