+++ src/dom/native/BaseTabNavigationElement.ts
@@ -2,12 +2,12 @@ import { ViewNode, logger as log } from "../basicdom";
 import { TabNavigationBase } from 'tns-core-modules/ui/tab-navigation-base/tab-navigation-base'
 import { TabStrip } from 'tns-core-modules/ui/tab-navigation-base/tab-strip';
 import { TabContentItem } from 'tns-core-modules/ui/tab-navigation-base/tab-content-item';
-import NativeElementNode from "./NativeElementNode";
+import NativeViewElementNode from "./NativeViewElementNode";
 
-export default class BaseTabNavigationElement extends NativeElementNode {
+export default class BaseTabNavigationElement extends NativeViewElementNode<TabNavigationBase> {
 
-    constructor(tagName: string, viewClass: typeof TabNavigationBase) {
-        super(tagName, viewClass, null);
+    constructor(tagName: string, viewClass: new () => TabNavigationBase) {
+        super(tagName, viewClass);
     }
 
     get nativeView(): TabNavigationBase {
@@ -20,12 +20,12 @@ export default class BaseTabNavigationElement extends NativeElementNode {
 
     onInsertedChild(childNode: ViewNode, index: number) {
         try {
-            if (childNode instanceof NativeElementNode && childNode.nativeView instanceof TabStrip) {
+            if (childNode instanceof NativeViewElementNode && childNode.nativeView instanceof TabStrip) {
                 log.debug(`adding tab strip to nav`);
                 this.nativeView.tabStrip = childNode.nativeView;
             }
 
-            if (childNode instanceof NativeElementNode && childNode.nativeView instanceof TabContentItem) {
+            if (childNode instanceof NativeViewElementNode && childNode.nativeView instanceof TabContentItem) {
                 log.debug(`adding tab content to nav`);
                 let item = childNode.nativeView;
                 //wait for next turn so that any content for our tab is attached to the dom
@@ -43,12 +43,12 @@ export default class BaseTabNavigationElement extends NativeElementNode {
 
     onRemovedChild(childNode: ViewNode) {
         try {
-            if (childNode instanceof NativeElementNode && childNode.nativeView instanceof TabStrip) {
+            if (childNode instanceof NativeViewElementNode && childNode.nativeView instanceof TabStrip) {
                 log.debug(`removing tab strip from nav`);
                 this.nativeView.tabStrip = null;
             }
 
-            if (childNode instanceof NativeElementNode && childNode.nativeView instanceof TabContentItem) {
+            if (childNode instanceof NativeViewElementNode && childNode.nativeView instanceof TabContentItem) {
                 log.debug(`removing content item from nav`);
                 let items = (this.nativeView.items || []).filter(i => i != childNode.nativeView);
                 this.nativeView.items = [];
