+++ src/dom/native/FrameElement.ts
@@ -1,11 +1,12 @@
 import { createElement, ViewNode, logger as log } from "../basicdom";
-import { Frame } from 'tns-core-modules/ui/frame'
-import NativeElementNode from "./NativeElementNode";
+import { Frame, View } from 'tns-core-modules/ui/frame'
 import PageElement from "./PageElement";
-export default class FrameElement extends NativeElementNode {
+import NativeViewElementNode from "./NativeViewElementNode";
+
+export default class FrameElement extends NativeViewElementNode<Frame> {
 
     constructor() {
-        super('frame', Frame, null);
+        super('frame', Frame);
     }
 
     setAttribute(key: string, value: any) {
@@ -13,17 +14,10 @@ export default class FrameElement extends NativeElementNode {
             log.debug(`loading page ${value}`);
             let dummy = createElement('fragment');
             let page = new (value as any)({ target: dummy, props: {} });
-            (this.nativeView as Frame).navigate({ create: () => (dummy.firstElement() as NativeElementNode).nativeView });
+            (this.nativeView as Frame).navigate({ create: () => (dummy.firstElement() as NativeViewElementNode<View>).nativeView });
         }
     }
 
-    get nativeView(): Frame {
-        return super.nativeView as Frame
-    }
-
-    set nativeView(view: Frame) {
-        super.nativeView = view
-    }
 
     //In regular native script, Frame elements aren't meant to have children, we instead allow it to have one.. a page.. as a convenience
     // and set the instance as the default page by navigating to it.
