+++ src/dom/native/PageElement.ts
@@ -1,17 +1,9 @@
 import { Page } from 'tns-core-modules/ui/page'
-import NativeElementNode from "./NativeElementNode";
+import NativeViewElementNode from './NativeViewElementNode';
 
 
-export default class PageElement extends NativeElementNode {
+export default class PageElement extends NativeViewElementNode<Page> {
     constructor() {
-        super('page', Page, null);
-    }
-
-    get nativeView(): Page {
-        return super.nativeView as Page
-    }
-
-    set nativeView(view: Page) {
-        super.nativeView = view
+        super('page', Page);
     }
 }
